refactored tags implementation
