+++ components/Tags.jsx
@@ -1,25 +1,26 @@
-import React from "react";
+import React, { useState } from "react";
+import Tag from "./Tag";
+// import { tagsList } from "@/public/data/tagsList";
 
 const Tags = () => {
+  const tagsList = ["Politics", "Music", "Culture", "School", "Random"];
+  const [selectedTag, setSelectedTag] = useState("");
+
+  const handleTagClick = (tagName) => {
+    setSelectedTag(tagName);
+  };
+  console.log(selectedTag);
   return (
     <div className="flex items-center gap-x-5 mb-10">
-      {/* <div className="flex items-center gap-x-5 mb-10"> */}
       <p className="text-white font-semibold text-2xl">Tags:</p>
-      <p className="bg-white text-black p-1 px-4 rounded-3xl font-medium text-lg">
-        Culture
-      </p>
-      <p className="bg-white text-black p-1 px-4 rounded-3xl font-medium text-lg">
-        Music
-      </p>
-      <p className="bg-white  text-black p-1 px-4 rounded-3xl font-medium text-lg">
-        School
-      </p>
-      <p className="bg-white text-black p-1 px-4 rounded-3xl font-medium text-lg">
-        Politics
-      </p>
-      <p className="bg-white text-black p-1 px-4 rounded-3xl font-medium text-lg">
-        Random
-      </p>
+      {tagsList.map((tag, index) => (
+        <Tag
+          key={index}
+          name={tag}
+          isSelected={selectedTag === tag}
+          handleSelected={handleTagClick}
+        />
+      ))}
     </div>
   );
 };
